inseptor

+++ src/store/actions/axiosInstance.ts
@@ -19,29 +19,29 @@ axiosInstance.interceptors.response.use(
   async (error) => {
     const originalRequest = error.config;
 
-    
+
     if (error.response?.status === 401) {
 
       if (originalRequest._retry) {
-      
-        localStorage.clear();  
-        store.dispatch(logout()); 
-        return Promise.reject(error);  
+
+        localStorage.clear();
+        store.dispatch(logout());
+        return Promise.reject(error);
       }
 
- 
+
       originalRequest._retry = true;
 
       const refreshToken = localStorage.getItem("refreshToken");
 
       if (!refreshToken) {
-   
+
         store.dispatch(logout());
-        return Promise.reject(error);  
+        return Promise.reject(error);
       }
 
       try {
-    
+
         const response = await axiosInstance.post(
           "/RefreshToken",
           { refreshToken },
@@ -52,26 +52,37 @@ axiosInstance.interceptors.response.use(
 
         const { accessToken, refreshToken: newRefreshToken } = response.data.data;
 
-      
+
         localStorage.setItem("token", accessToken);
         localStorage.setItem("refreshToken", newRefreshToken);
 
-
+         if (error.response.status == 401 ){
+          localStorage.clear();
+          store.dispatch(logout());
+        }
         axiosInstance.defaults.headers.common["Authorization"] = `Bearer ${accessToken}`;
         originalRequest.headers["Authorization"] = `Bearer ${accessToken}`;
 
         return axiosInstance(originalRequest);
       } catch (err) {
+
        
-        localStorage.clear();  
-        store.dispatch(logout()); 
-        return Promise.reject(err);  
+        return Promise.reject(err);
       }
     }
 
-    return Promise.reject(error);
+    throw error;
   }
 );
 
 
 export default axiosInstance;
+
+
+
+
+
+
+
+
+
